requirements.txt
torch>=2.2.0
transformers>=4.40.0
accelerate
bitsandbytes
datasets
numpy
pandas
matplotlib
tqdm
pyyaml
